Return ResultType.SKIPPED for is_skip
Currently is_skip() returns True or False to
determine if the step needs to be skipped or
not. However there are cases where some value
needs to be return even when the step are skipped
so that the step can be made idempotent.

Use ResultType.SKIPPED as return value for
is_skip().
Refactor plan execution code to a single function
as it is repeating for every cli command.