Add configure step
This change adds a configure step that can be run after bootstrap.
The configure step will optionally setup default images, flavors
etc in the cloud as well as configuring networking.

The last step in the plan configure executes runs an action on the
openstack-hypervisor charm to set unit specific values such as
which nic to use for North/South traffic. This change also adds
that step to the node join command.

Depends-On: https://review.opendev.org/c/openstack/charm-openstack-hypervisor/+/881493