Change Deployment status logic to ignore some apps
Currently once the openstack is deployed via terraform,
sunbeam waits for all the apps to be active and idle.
However in cases where storage node is not added yet,
cinder-ceph application will be in blocked state.
Hence wait for the apps logic need to be changed to
ignore some apps as and when required.

If there are storage nodes, cinder-ceph is removed
from the list of apps that need to be waited for
active state.