Update controller endpoint when DB has empty list
When user inputs wrong management_cidr, the controller
api endpoints are updated with empty list in cluster
database. Rerunning the bootstrap with proper
management cidr only reads the endpoints from the
cluster DB which is empty list and further connectivity
to juju is broken.
Do not skip ClusterUpdateJujuController Step if
api endpoints are empty.