Enhancements to node removal
Removing a node is currently broken as the sunbeam and hypervisor
charm remove steps are not being called. This change adds them in.

For the hypervisor removal this change adds a check to see if any
guests are running on the hypervisor targeted for removal. If there
are then an error is thrown unless the "--force" flag has been set.

As part of the hypervisor removal the compute and network agents are
also de-registered.