bootstrap: implicit control role
Bootstrapping without the control role does not make sense so
make it implicit and don't allow users to drop it - only additional
roles can be provided.

Additional nodes joining the cluster don't have to be control nodes
so its safe to continue to have this as an option for the 'cluster join'
command.