bootstrap: support with just control role
Ensure that all charms are deployed with units added to each based
on the roles provided during bootstrap.

This allows a control plane only node to be create supporting a
dedicated set of control plane nodes in a deployment.

In order to support this topology update the configure process
to only attempt to configure the local hypervisor networking if
the node has the compute role.