Add machine-ids as part of terraformvars to clusterdb
Currently machine id's are not saved as part of terraformvars
in clusterdb for any of the machine charms deployed. Write
terraformvars for microk8s, microceph, sunbeam-charm in clusterdb.

Move the common functionality in Deploy application, Add unit,
Remove unit to a base class so that the commands microk8s, microceph
and sunbeam-machine can derive from common functionality classes.