resize: misc fixes to resize of control plane
When a deploy is bootstrapped with a dedicated control node, ceph
is not enabled as part of the control plane deployment.

Storage nodes (and associated ceph units) may have been added during
the deployment of the rest of the nodes so re-assess as to whether
ceph integrations should be enabled during resize operations.

Ensure that HA services (mysql, ovn-central) scale with the
number of control nodes and only when there is sufficient to
jump from 1->3 to ensure quorum can be achieved by the specific
applications.

Add test cases to cover all of the scale helpers.

LP:#2043083