Generalise terraform functions into manifest
* All the commands have logic to copy the teraform
plans to snap user common and creation of
terraform helper object. Move this logic as part
of Manifest class.
* All the plans more or less get terraform vars
from cluster db, update with generated tfvars based
on plan, update with tfvars from manifest, write
tfvars to the terraform plan and apply the plan.
Move this logic as part of Manifest class