Add TLS CA plugin
Add TLS CA plugin to enable manual-tls-certificate
operator. The plugin takes ca certificate and chain
as arguments. Endpoints can be specified to integrate
operator with traefik-public or traefik internal.

Add command configure to get the certificate from
user and update certificates in the traefik application.

Transfer the CA and chain certificate to keystone so
that the certs can be distributed.