Bump the HYPERVISOR_UNIT_TIMEOUT
A fixed timeout is not good in general without checking the status if a
charm is really in the error/blocked status and stuck without a human
intervention or hooks are still going on and just taking time. However,
"Timed out while waiting for units openstack-hypervisor/N to be ready"
is the most frequent timeout I saw in the testing so bump it not to hit
into it with normal circumstances.

Closes-Bug: LP: #2065866